requirements.txt
django
django-versatileimagefield
psycopg2
django-admin-sortable2
django-debug-toolbar
uwsgi
pytz
django-admin-tools
mailchimp
stripe